requirements.txt
cloudscraper
discord
python-dotenv
bs4
